backend/requirements.txt
# Framework
fastapi==0.120.4
Flask==3.1.2

# ASGI Server (standard includes watchfiles, httptools, etc.)
uvicorn==0.38.0

# Database (Async ORM)
SQLAlchemy==2.0.44
asyncpg==0.30.0
psycopg2-binary==2.9.11

# Config & Validation
pydantic==2.12.3
email-validator==2.3.0
pydantic-settings==2.11.0 # For BaseSettings in Pydantic V2+
python-dotenv==1.2.1

# Security (Password Hashing & JWT)
passlib==1.7.4
argon2-cffi==25.1.0
python-jose==3.5.0
cryptography==46.0.3
python-multipart==0.0.20

# Migrations
alembic==1.17.1

# File & Utils
aiofiles==25.1.0
Markdown==3.9

# Networking / HTTP Clients
httpx==0.28.1
requests==2.32.5

# Testing
pytest==9.0.1